refactor: 음악 인증 관련 비즈니스 로직 packaging
