+++ src/main/java/io/github/eappezo/soundary/services/music/application/authentication/MusicPlatformAuthenticationCacheAdvice.java
@@ -1,4 +1,4 @@
-package io.github.eappezo.soundary.services.music.application;
+package io.github.eappezo.soundary.services.music.application.authentication;
 
 import io.github.eappezo.soundary.core.cache.CacheAdvice;
