+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/gateway/SpotifyOAuthGateway.java
@@ -3,8 +3,8 @@ package io.github.eappezo.soundary.services.music.infrastructure.gateway;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import io.github.eappezo.soundary.services.authentication.domain.exception.AuthenticationFailedException;
-import io.github.eappezo.soundary.services.music.application.MusicPlatformOAuthGateway;
 import io.github.eappezo.soundary.services.music.application.PlatformAccessToken;
+import io.github.eappezo.soundary.services.music.application.authentication.MusicPlatformOAuthGateway;
 import io.github.eappezo.soundary.services.music.domain.MusicPlatform;
 import io.github.eappezo.soundary.services.music.infrastructure.MusicPlatformOAuthConfig;
 import io.github.eappezo.soundary.services.music.infrastructure.MusicPlatformOAuthProperties;
